:truck: Move module filestorage.adapters.{ => observers.}git
